+++ src/linux.rs
@@ -822,6 +822,10 @@ pub fn sc_system_update_rtools40() -> Result<(), Box<dyn Error>> {
     Ok(())
 }
 
+pub fn sc_system_rtools() -> Result<(), Box<dyn Error>> {
+    Ok(())
+}
+
 pub fn sc_rstudio_(version: Option<&str>, project: Option<&str>, arg: Option<&OsStr>)
                    -> Result<(), Box<dyn Error>> {