+++ src/macos.rs
@@ -790,9 +790,13 @@ pub fn sc_system_update_rtools40() -> Result<(), Box<dyn Error>> {
     Ok(())
 }
 
+pub fn sc_system_rtools() -> Result<(), Box<dyn Error>> {
+    Ok(())
+}
+
 pub fn sc_system_detect_platform(_args: &ArgMatches, _mainargs: &ArgMatches)
                                  -> Result<(), Box<dyn Error>> {
-    // Nothing to do on Windows
+    // Nothing to do on macOS
     Ok(())
 }